Delete unpublishings in ScheduledEditionPublisher
If a document is unpublished and then scheduled-published, whitehall
gets into a weird state where the document is reported as both
published and unpublished.  This is because it doesn't delete the
unpublishing record.

This commit copies the code to delete an unpublishing from the
EditionPublisher.  This makes the fire_transition! for scheduled
editions almost the same as the fire_transition! non-scheduled
editions, but with one important difference: a scheduled edition isn't
validated.

This is intentional behaviour from the first implementation, 19fca44:

> Scheduled editions should bypass validation when being published to
> ensure that a change in the validation rules inbetween being
> scheduled and the scheduled publication date doesn't prevent the
> edition being published.

It would be nice to delete ScheduledEditionPublisher.fire_transition!,
but not really if we have to complicate EditionPublisher.fire_transition!
with selective validation logic.