Remove a load of old reporting tasks
These tasks were added several years ago, and don't seem to be
documented/used anywhere. Since the work to fix the linting issue
(for the top-level method) is quite tedious, it seems better to
remove the older tasks that use it.

Alternative approach: just disable the cop for this one use case.