Fix subtle 'delegate' v Forwardable conflict
These two methods of delegation aren't compatible. Since 'delegate'
is the one preferred by RuboCop, this removes all instances of the
'Forwardable' module, and replaces them with plain 'delegate' calls.

Note that, for uses of 'def_delegator' (vs. 'def_delegator**s**'),
the third argument is actually an alias - not a delegated method.