Add custom error when updating an asset without any attributes
Whitehall's AssetManager::AssetUpdater is sending requests with apparently
empty `new_attributes` which is causing the AssetManager app to respond
with 'File must not be blank' errors.

The intention is that empty attributes should not be sent to AssetManager,
so this commit adds an explicit error message to catch these before the request is made.