Present internal taxon name to publishers
This will help give more context to areas of the taxonomy that are mixed,
such as the new coronavirus branch.  This has areas that collide with others
(such as education or business topics), so additional information in the
publishing tool is useful.

We are working on patterns for GOV.UK frontend that help resolve this problem
for users.