Update change note form text
This changes the text and layout of the admin change box forms.

There is a need for the hint text and information for a major change
to appear between the form label and the text_area, rather than as placeholder text.

Currently this is not possible with the FormBuilder which can only render a label text
in bold.

Rather than extend the FormBuilder further, as the long-term intention should be to
move to govuk-publishing-components, this PR appends the necessary text via javascript.