Explain bug and link to support form
Sometimes a document can get into a state whereby its latest
edition is deleted, and the most recent non-deleted edition
is marked as 'superseded'. The result is a 'Go to most
recent edition' button that just refreshes the same page
again and again.

A 'proper fix' needs careful consideration. When an edition
is deleted, should the previous edition itself be changed
from 'superseded', or should it be cloned to retain history
(i.e. a timeline of ['superseded', 'deleted', 'draft'])?
Do we have any way of knowing what the previous edition
state was prior to being 'superseded' (e.g. 'draft' /
'published'), and therefore what it should now be set to?

In the absence of a permanent fix, a pragmatic stop-gap
would be to explain the situation to users and link to a
support form so that they can ask for help if needed. The
extra context should enable us to help them more quickly
too.