Delete Document when deleting Edition, if it has no other Editions
We have a bug in DocumentCollection whereby if you add a Document
to a DocumentCollectionGroup (thereby making a
DocumentCollectionGroupMembership), then put that Document in a
'deleted' state... it no longer appears in the group, but the
association is still there, and the publisher is unable to delete
that group. They see the following error message:

"You can’t delete a group that has documents in it – remove the
documents first, or move them into another group."

A Document can get into a deleted state by, for example, being a
newly created Document which has just one Edition, which is in a
draft state. The publisher is able to 'discard' that draft, which
puts the Edition in a 'deleted' state, and the Document no longer
has any non-deleted editions, so it is 'deleted'.

I say 'deleted' with air-quotes because it isn't actually deleted;
its record still exists if you do an explicit `Document.find(THE_ID)`.
The same applies to the Edition - it exists but has a state of
'deleted'. If you do a generic ActiveRecord search, it won't appear
in the results, as [by default we scope our queries to filter out
'deleted' editions](https://github.com/alphagov/whitehall/blob/9dcec7b76ed6617a293bf51fdca372b1abd085e1/app/models/edition/workflow.rb#L21).

Because the Document doesn't _actually_ get deleted, its
[`delete_all` callback](https://github.com/alphagov/whitehall/blob/41b611f3d8c092706b3a27f7063b6d014a0efb88/app/models/document.rb#L38)
never gets fired, so its associated document collection group
memberships persist.

This commit adds an `after_destroy` callback to the Edition model,
to check if its associated Document has any other Editions associated
with it. If it doesn't, it can be safely deleted - which will
automatically trigger the cleaning up of any other associations.

Trello: https://trello.com/c/579DqsRD/1976-8-investigate-why-sometimes-document-collections-have-a-bad-state