Simplify deleting document collection groups
Rather than having the user delete the group items if there are any,
just handle that as part of deleting the group.

This also resolves an issue where a document associated with the group
with just one edition, which happens to be deleted, would prevent the
group from being deleted, as it's not visible to the user to remove
from the group.