Publish history pages as history pages
This is a stepping-stone to having the content in government-frontend:
get the document type in the publishing-api right, and also get the
content in the right search index.

This isn't strictly necessary, but it's nice to see that that part of
things all works before switching the publishing and rendering app
over.