Guard against deleting unpublished drafts in the EditionDeleter
This is the same constraint as the user interface. At least at the
moment, if you delete a edition which is in the draft state, because
it was unpublished, you end up a bit stuck, not being able to create a
new draft.

This doesn't fix that issue, but maybe this will help avoid documents
getting stuck like this in the future.