Use 'update!' to surface save errors
Previously we switched from 'update_attribute' to 'update' to fix
linting issues, but this carries a risk of silently failing to save
changes, since 'update' will run validations. This switches all uses
of 'update' to 'update!' to ensure we become aware of any issues in
case they are not covered by the tests.

Note that some method calls cannot be changed without further work,
so this disables the Cop instead:

- Admin::WorldwideOrganisationsController#create for some reason
returns a 200 response if the creation failed, and relies on the
view to find and show any errors.

- StatisticsAnnouncement#cancel! uses 'save' to return the boolean
result of the operation.