Explicitly return only the fields that are used
We were seeing errors in Logit as follows:

```
[4938944e-2f85-4a07-902f-585841876d22] Marshalling error for key 'count=1000/fields=["display_type", "title", "link", "public_timestamp", "format", "content_store_document_type", "description", "content_id", "organisations", "document_collections"]/filter_topical_events=coronavi:md5:a392a8c87eaf0fcfcbf88dcf77d910aa': no _dump_data is defined for class Thread::Mutex
```

The fields correspond to the `Rails.cache.fetch` call in
SearchRummagerService. Following the request ID, we can find other
logs that shed more light:

```
[4938944e-2f85-4a07-902f-585841876d22] You are trying to cache a Ruby object which cannot be serialized to memcached.
```

It appears what we are caching is an instance of 'GdsApi::Search'
(see rummager.rb: `Whitehall.search_client = GdsApi::Search.new`)
which evidently is not serializable.

Caching only a hash of what we need seems to do the trick to fix
it. In most places, 'results' is the only field used, but in the
'app/views/topical_events/show.html.erb' we can see that 'total'
is used too. I can't see any other fields used.

Trello: https://trello.com/c/ZZCFTp52/1974-5-investigate-whitehall-marshalling-errors