Prevent updating asset if missing asset file information
We're seeing [issues in Sentry][sentry-issue] as follows:

> URL: https://asset-manager.production.govuk-internal.digital/assets/5b115d34ed915d2cc681ab6f
> Response body:
> {"_response_info":{"status":["File can't be blank"]}}

An attempt was made in [#5581][log-attempt] to catch this error
and provide more detail to help make the error actionable.
Unfortunately, the error condition coded in that PR doesn't
ever appear to happen, so we're still seeing the more obtuse
Sentry logs.

It is evident that `new_attributes` ISN'T empty, but the file
is considered 'blank' when attempting to send to Asset Manager.
This commit has an explicit check that the `attachment_data.file`
property exists before trying to update the asset.

This is a short/medium term approach which will give us a slightly
more meaningful error message, and prevent us from making the
faulty Asset Manager call in the first place. In the long term
we will still need to figure out how some assets get into this
state and prevent that from happening.

Trello: https://trello.com/c/FDu35RHS/1977-investigate-whitehall-sending-blank-file-when-updating-assets

[log-attempt]: https://github.com/alphagov/whitehall/pull/5581
[sentry-issue]: https://sentry.io/organizations/govuk/issues/1650510012/?project=202259&query=blank&statsPeriod=14d