Serve whitehall assets without a host by default
This change is made in order to implement [RFC 115][1] where the static
application assets are moved to be hosted on www.gov.uk rather than
assets.publishing.service.gov.uk.

Applying this to Whitehall is quite complex due to it's role as both a
publishing and rendering application.

In this change I've set an assets prefix in config/application.rb of
assets/whitehall. This is to replace the previous inconsistent one of
/governemnt/assets - I presume this was set-up at a time when
/government proxied all requests to Whitehall.

I've removed the proc that is used to determine asset path in the
production configuration. As far as I can tell this is something that is
no-longer necessary since asset-manager began serving Whitehall uploads.

I've configured the means to set an asset host by an env var of
ASSET_HOST. This is likely to only be needed in development environments
where the root hostname isn't set-up to proxy assets to the Whitehall
application.

[1]: https://github.com/alphagov/govuk-rfcs/blob/master/rfc-115-enabling-http2-on-govuk.md