Use Whitehall.public_root for absolute asset urls
With the switch in hostname of assets these assets as no longer served
from the public_asset_host. In every normal case we shouldn't need a
host for assets as they are served relative to a request, these assets
however represent an exception. These are bizarre and dangerous cases
where a fingerprinted Whitehall asset is put into a content item.

This approach is a really risky idea as if the asset changes at all the
fingerprint will change and links to the asset may stop working.
Whitehall is the only application that does this and it really should
stop.

These assets likely won't work in a development environment where a
developer hasn't set up a proxy on their router app from
assets/whitehall to the whitehall app. For anyone who is considering
fixing that I would recommend that you instead spend the time trying to
stop Whitehall doing this terrible thing with it's asset URLs.