Isolate placeholder image urls to LeadImagePresenterHelper
We only need to apply the asset pipeline image_url function to the
placeholder image as that is the only asset that is within asset
pipeline. By moving this code we only use it when it's needed and avoid
replicating it.

It is, of course, not good that we create these asset fingerprinted URLs
and store them in content items. They are at risk of breaking any time
the asset is edited. I feel that this change benefits this oddity
somewhat by at least moving it all into the same place.