Create a new presenter for the ministers index static page.
This work is part of the work to migrate rendering of this page from whitehall
to collections, the first step of which is to have a content item presented to
the publishing-api that can in future be rendered by collections.

I have also added necessary tests to check that the content item contains what
it should do, and also that it abides by the schema set out in govuk-content-schemas.
(https://github.com/alphagov/govuk-content-schemas/commit/66ebaf0e2d0454719c6f3a96f1b5ae22317a50a4#L9-L10)

https://trello.com/c/I92GWawi/2027-5-write-ministers-index-presenter