Make attachment thumbnails use public GOV.UK hostname
Attachment thumbnails are inconsistent with the other means that assets
can be embedded in content items. Unlike placeholder images and
fractions these images are not embedded with a hostname. The change here
makes this approach consistent with the other two.

As far as I can tell these URLs should have been set an absolute URL
with the asset host around the time https://github.com/alphagov/whitehall/pull/2481
was introduced. In theory these should have stopped working when
Whitehall started using the assets hostname for serving assets, however
the /government content item prefix route provided a loop hole that kept
them working.

I considered whether to either make the other URLs used to embed assets
relative so that all of these use less data and are a bit more flexible,
however this approach conflicts with some JSON schema semantics. The
placeholder image is used with a URI schema validation [1] and as per
JSON Schema a URI should be absolute with a scheme [2]. Thus the
consistent approach was setting these all as full URIs.

It's worth noting (as I have in many related commits) that we really
shouldn't be embedding Whitehall assets in content items as this a
rather fragile abstraction and a strange coupling of an applications
internal structure with public content.

[1]: https://github.com/alphagov/govuk-content-schemas/blob/91ef588841f6073b3cc0d362f562ed298a4c0a15/dist/formats/speech/publisher_v2/schema.json#L338
[2]: https://json-schema.org/draft-06/json-schema-release-notes.html#formats-uri-vs-uri-reference