Upload csv exports to s3 and allow requesting them, as with specialist-publisher
This uses a similar approach to that in specialist-publisher
(alphagov/specialist-publisher#1576 and alphagov/specialist-publisher#1625):
it adds a new controller which retrieves the file from S3 and proxies it
to the user. Doing this instead of simply sending the user a link to the
S3 bucket allows us to use Whitehall's access control and to avoid
making the bucket itself publicly-readable.

https://trello.com/c/5xKHcam8/2012-5-whitehall-replace-attachments-in-documentlistexport-so-that-notify-can-be-used