Fix some incorrectly redirected HTML attachments
This addresses some mistakes I made when fixing a wider issue
regarding redirects for non-en content [1][2].

1: https://trello.com/c/1Xd84A0k/14-8-fix-redirects-that-havent-been-created-correctly-for-non-english-content
2: https://github.com/alphagov/publishing-api/pull/1736/commits/7903e65c2b32ed363f47994287c8896862924165

There were many (~7000) draft redirects that weren't published, and
were published as part of [1]. However, some of those redirects
shouldn't have been published, as the target of the redirect wasn't
published yet.

This rake task fixes the affected pages by having Whitehall publish
the HTML attachment again, replacing the redirect. To guard against
the situation changing between when this is written/merged and run on
Production, the task checks if the path is in the broken state before
making the change (it's not a perfect check, but hopefully good
enough).