Make the images part of the edit edition form more readable
I'm not sure what's gone wrong here, there are some classes like
image-container that are in the HTML, but there's no corresponding
CSS. The CSS for image-container seems to have been removed 8 years
ago back in 2012 [1].

1: cb59ac01728238cee9d2a938d45b9a635d96efe4

These changes don't try and make things better, just less worse. Each
image will take up less vertical space on the page, and it should be
visually clearer which image the form elements relate to.

I was looking at this because I wanted to see if something could be
done here to clear up confusion about how the special images for
speeches are handled, but that's hard to do with the interface in such
a state...