Fail validation if URL contains whitespace
A user managed to save a URL with whitespace which lead to
publishing-api rejecting the unpublishing request but whitehall saving
the consolidated URL leaving the edition is a funky state. This PR
prevents this from happening in the future.

This also switches over to using ruby's URI instead of Addressable. This
seems to of been added originally due to limitations in our URL
validation however from my testing using plain old ruby's URI is fine.
[Original PR](https://github.com/alphagov/whitehall/commit/5b313517fec1b53271ada7214b3a30dea56a97e6)

Now the class validating the URL matches what is present in
[publishing-api](https://github.com/alphagov/publishing-api/blob/93605514d7c32ddcb93bf793529de00b64cc1c39/app/validators/routes_and_redirects_validator.rb#L171)
so we should no longer get a disparity in validating URL's between the
two apps.

Zendesk:
https://govuk.zendesk.com/agent/tickets/4144535