Republish organisations when roles are changed
In the publishing API organisation presenter, we link to people,
instead of roles:

    def people_content_ids(role:)
      item.send("#{role}_roles")
        .order("organisation_roles.ordering")
        .map(&:current_person)
        .compact
        .map(&:content_id)
    end

This means that we can't rely on link expansion in the publishing-api
itself to handle the case where we link or unlink a role to an
organisation, because the "link expansion" for that case is happening
in whitehall.

This is a quick fix to make updating roles work again, but ideally
we'd change the presenter to link to roles, not to people.