Remove image "placeholder" alt text
When the alt text field is left blank when uploading an image for a news
article, the word 'placeholder' is used as alt text instead.
This is a problem for screen reader users as the word doesn't make sense
in relation to the image.
It's better to leave alt text empty so screen readers can ignore it.