Ensure that a fresh stub is created for every scenario
The TaxonomyHelper module sets `Redis.current = stub` in order to not
have to pass the stubbed client into RedisCacheAdaptor but this leads
to the stub leaking between scenarios. This currently generates a
warning but will in a future version of Mocha raise a Mocha::StubbingError
exception so fix this by resetting `Redis.current = nil` before every
scenario. We can't set the stub directly as Mocha complains about
the method be used outside of a test/scenario.