Allow FCDO to tag to worldwide taxonomy
This adds FCDO's `content_id` to the list of `content_id`s of
organisations allowed to tag their documents to the worldwide taxonomy.

This is part of the work around the Machinery of Government changes that
will see FCO and DFID merge to become a new department.

Trello card: https://trello.com/c/Meb6kRZu/2099-1-allow-fcdo-to-tag-to-worldwide-taxonomy