all_nation_applicabilty checkbox is empty on first creation
When first creating the first edition of a new document, the "applies to
all UK nations" we do not want the check box to be auto-filled.

We assert this by checking that the Edition
does not have an associated Document.

Once the user has attempted to save the Edition we can use the
Edition's own `all_nation_applicability` attribute as this has been
confirmed by the user.