Add seniority to ministerial roles
Part of the work to migrate the rendering of /government/ministers
from whitehall to collections. Seniority will be used by collections
to sort the list of ministers so needed to be added to the content
item to collections had this information via link expansion.

More information here: https://trello.com/c/T6u4LjxS/2062-3-add-sorting-information-to-ministers-content-item

and here: https://trello.com/c/loAlkIp7/1715-8-migrate-government-ministers-to-collections

Relevant PR's:
https://github.com/alphagov/govuk-content-schemas/pull/999
https://github.com/alphagov/publishing-api/pull/1843