Set all_nation_applicability for withdrawn editions
In 93acb17, a migration was added which updates the
`all_nations_applicability` field for editions which have any nation
applicability specified.

However, this migration did not make the change for withdrawn
publications, which had the value set as `true` even if applicable
nations are specified.  This led to a validation error when publishers
attempt to unwithdraw the edition.

Therefore adding another migration to update this value for withdrawn
editions.