Request accessible format toggle link correction
The document list has a link that shows/hides a bit of text.
The link reads "Request an accessible format" and toggles an explanation
of how to get in touch to ask for a different document format.
This is styled as a blue underlined link but has a role of button.
People who use assistive tech may have trouble determining what this
element does.

Change the "Request an accessible format" toggle to use a details
component instead. The details component uses details/summary, which is
more accessible.