Add a migration
Add a migration to add a string value to edition for for the
image_display_type. With the possible values being "organisation_image",
"no_image" or "custom_image".
(we considered using a boolean however when we tested it out we could
not set the value to nil with the radio buttons since it was treated as
false)