Fix stray p tag issue
It has been flagged that there are stray p tags on pages with accessible
format requests. It seems like this is happening because the details
component is contained inside the paragraph, when the paragraph should
really only be wrapping the intro text.

This moves the details component out of the paragraph.