Add lang="en" to publication dates
Publication time/date seems to always be in English regardless of
current locale. As we are missing the config locales to display dates in
every language, it makes sense to add lang="en" to the time element.

This ensures that screen readers might read would read that text
correctly, even though it is in a different language than the main
document.