Add lang to page contents title when untranslated
In some instances, on translated world location pages, the "Contents"
title will be in English and not marked as such using an appropriate
lang attribute.

This adds lang="en" where "Contents" is in English, to ensure that
screen readers read this word in an understandable fashion.

It also passes through the translated title, if a translation exists
(for instance in French the translation is "Contenu"). Therefore when
viewing the French translated version of a page, the contents list title
should be "Contenu", and should not have a lang attribute.