On Edition.primary_locale change, update the translations
Only in the case where an edition has a single translation, keep that
translations locale in sync with the edition primary locale. This
handles the case where the primary locale of the edition is
changed. Rather than no translation existing for the primary locale at
this point, change the translation locale to match the primary locale.

This is primarily aimed at allowing editions with the wrong primary
locale to be corrected, so it definitely makes sense in that scenario.