Write a failing test reproducing primary_locale translations issue
When a publisher changes the primary locale of a world news story,
we end up in a strange state whereby the draft edition has two
translations: one with the old locale, and one with the new.
When the publisher attempts to publish the draft, Publishing API
raises an error: "Cannot publish an already published edition".

This happens because both translations are sent to publishing API,
and the translation with the old locale has already been published,
so cannot be published again. The simplest solution would be to
remove the redundant translation, so that it is automatically
[removed at time of publish]
(https://github.com/alphagov/whitehall/blob/277d3216a1009b69cd04b63a1185a7b19e7a4c26/lib/translatable_model.rb#L29-L31).

Although [we generally try to avoid writing cucumber tests]
(https://docs.publishing.service.gov.uk/manual/conventions-for-rails-applications.html#featuresystem-testing),
it was the only way I was able to reproduce Whitehall creating
two translations for the draft edition. At the unit test level it
seemed to produce just one translation as expected, and at the
integration test level I had difficulty stubbing all the necessary
Publishing API associations (particularly 'world locations', which
must be tagged to the edition in order for it to be publishable).

`edition.government` was returning `nil`, so _standard_fields.html
was throwing a "Cannot call `.name` on NilClass" error. As this is
an admin-facing view and is only offered as 'extra' info in the
label creation, I didn't see a problem with using the Safe
Navigation Operator to silently capture the error (which would be
pretty evident in the UI itself if it happened). It would be
better than the app crashing, which is what would happen at the
moment in the event of an associated Government not being found.

Trello: https://trello.com/c/y9H3BLeG/2168-new-editions-of-foreign-language-only-world-news-stories-cant-be-published