Fix World News Story translations publishing issue
See previous commit. When a world news story's primary_locale is
changed from `en` to another language, we end up with two
translations: one is the old `en` translation and the other is the
new locale. We need to remove the old translation so that the new
one can be published and the English version removed.

There is a warning on the 'Edit news article' UI:
"Warning: News stories without an English version cannot have
other translations."

This commit goes a little further than that, in that if the
English version is not the _primary_ locale, the other translations
are removed. Unfortunately I can't just check if the edition has
an English version, as the fact that it _does_ is the bug I'm
trying to fix!

https://trello.com/c/y9H3BLeG/2168-new-editions-of-foreign-language-only-world-news-stories-cant-be-published