Disable scss/at-extend-no-missing-placeholder for organisations.scss
I spent a good while trying to work out if this could be swapped around
to use placeholders or not. I found a bunch of the root classes aren't
used (such as organisation-logo-identity, organisation-logo-identity-medium,
and organisation-logo-identity-large) and are reasonable candidates.
However after initial progress I eventually concluded that these classes
were all coupled so tightly that it would be very hard to achieve this
without breaking something - and I'm really unsure how to tell if I've
broken anything given the huge variety of classes at play.

It's very hard in particular to work out which classes are used in the
application as they are generated automatically by a helper function
(app/helpers/logo_helper.rb) so even searching for which classes are in
use or not is very difficult.