Fix linting (and logic) issue in tabs js helper
This file had camel case and unused variable offences, it seems that
this previously had a mistake when setting the scroll position to use
both y axis rather than x and y axis. I assume no-one caught this as the
X axis is rarely scrollable.

I've put this in a separate commit to try help me remember to test this
individually.