Ignore no-duplicate-selectors on base.scss
The #whitehall-wrapper selector is defined multiple times in this
stylesheet, typically under a heading to group various imports together.

The way to fix this would be to include all the imports within a single
selector of #whitehall-wrapper. I could see the way to do this would
involve moving the `@import "layouts/main"` import at the top and then
combining the other imports into a single definition of #whitehall-wrapper.
However I was nervous about the consequences of doing this as I have no
idea if any CSS relies on this ordering.