Split out 'call' method into two separate methods
I note that the interface was deliberately unified in June
2018 to this single method `call`:
https://github.com/alphagov/whitehall/commit/df25cdb5d6266a5c833ce211e4751a1c19961b86

Back then the `AttachmentUpdater` class was a lot bigger and
benefited from a single entry point. Now that the class is so
small, we can afford to name the entry points better, and to
have simpler code.

The code in Whitehall only ever passes one of `attachment_data`
or `attachable` - never both at the same time - so it seems
unnecessary to try to handle both in one method.
Moreover, as it is possible to pass both attachment_data and an
attachable which is associated with that same attachment_data,
this would trigger `AssetManagerAttachmentMetadataWorker` twice
for the same attachment data, which could cause buggy behaviour.

Having two explicit entrypoints means that buggy code is less
likely. It's cleaner than writing code to raise an exception if
two things are passed, etc.