Don't update attachment metadata for already-deleted attachments
We're seeing lots of `AssetManager::ServiceHelper::AssetNotFound`
errors in Sentry (here's an [example][]), originating from
`AssetManagerAttachmentMetadataWorker`. Looking at the assets in
question, their respective 'attachment' always seems to be
`deleted: true`, which explains why the asset cannot be found.
It also begs the question: why are we trying to apply any workers
to them in the first place?

When an attachment is deleted, it shouldn't be touched again.
Once this commit is merged, we expect to see these kinds of errors
reduce by half. We still have to apply a similar fix to errors
originating from `AssetManagerAttachmentRedirectUrlUpdateWorker`,
which account for about half of these kinds of errors.

Trello: https://trello.com/c/RethzUSW/2088-5-fix-assetmanagerservicehelperassetnotfound-errors-on-staging-integration

[example]: https://sentry.io/organizations/govuk/issues/1668643818/events/f1bf7ff83ede4822b9a781459d749760/?environment=staging&project=202259&query=error.type%3AAssetManager%3A%3AServiceHelper%3A%3AAssetNotFound+transaction%3ASidekiq%2FAssetManagerAttachmentMetadataWorker&statsPeriod=90d