Add new Rake task to allow republishing of Take Part pages
We've added a new ordering option to the schema for Take Part pages that
allows us to store the ordering of the pages in the Publishing API. This
rake task simply gives us the ability to republish the pages with that
information.

I think it may be worth keeping around after its initial use until we
complete migration work, in case of future needs.