Update email signup with new design
This makes the signup page consistent with the generic start page
for content item signups [1]. Note that Whitehall seems to rely on
custom styles for margins, but the other styles aren't needed.

Note that the "govuk-!-font-weight-bold" class is required due to the
unfortunate "font-weight: inherit" CSS reset that pervades this app [2].
It's out-of-scope of this work to explore removing these ancient resets.

[1]: https://github.com/alphagov/email-alert-frontend/pull/918
[2]: https://github.com/alphagov/whitehall/commit/7718a08793aa89c50ef78e5a51de98ffc670ee1f