Fix zoom overflow
The big numbers affected in this commit currently overflow into the surrounding
region when a user increases text size but doesn't otherwise zoom. This corrects
that by reducing the sizes of the big numbers as they run out of space based on
em sizes. This will not make them illegible as they are still enormous in
comparison to the copy text.