Sets string_for_slug to title for DocumentCollections
Currently Editions with non-English locales have their slug set to
the Document ID, via the `ensure_document_has_a_slug` callback in Document
https://github.com/alphagov/whitehall/blob/67f42b449e0738453f3201906fa5685b84fa5193/app/models/document.rb#L135

These non-English slugs are set to nil via the `string_for_slug` method in Identifiable
https://github.com/alphagov/whitehall/blob/cba922543fdff10fdd7090e0facc2514c5daa632/app/models/edition/identifiable.rb#L50

As we now wish DocumentCollections to be created with non-English primary locales, it is undesirable for the
slugs to be the id.

This PR ensures DocumentCollections will always use the title regardless of locale.