Add Optional Locale to DocumentCollection Route
This commit rewrites the DocumentCollection Route to allow
an optional locale parameter. The previous version was only
on #show, so there should be no issue with switching this from a
resource to a simple get.

This allows the DocumentCollectionPresenter to append a `.:locale`
to a non-:en DocumentCollection's base_path when sending to the
PublishingAPI.