Stop mixing strings and symbols in locales
This makes `Whitehall.available_locales` more consistent by only using
symbols and alphabetically ordering the list.

It also adds a test to check that the locals in
`Whitehall.available_locales` and the files in `config/locales` match.
This is useful so we don't accidentally forget to add locales somewhere.

Trello:
- https://trello.com/c/u1FYvrIZ/2241-5-allow-primarylocale-of-consultations-to-be-changed
- https://trello.com/c/3MrqbQae/2200-5-allow-primarylocale-of-documentcollections-to-be-changed