Specify if locales use Latin-script writing system
This adds a new entry to each of the supported locales. The new entry is
a key/value pair where the key is `latin_script?` and the value is a
boolean: `true` if the language's writing system is a Latin-script
alphabet, `false` otherwise. To set the boolean, the locale files have
been inspected and, when in doubt,
[this list](https://en.wikipedia.org/wiki/List_of_languages_by_writing_system#Latin_script)
has been consulted.

For the reason `latin_script?` was chosen over `latin_alphabet?` or to
delve a bit deeper, see:

- https://en.wikipedia.org/wiki/Writing_system
- https://en.wikipedia.org/wiki/Latin_alphabet
- https://en.wikipedia.org/wiki/Latin_script
- https://en.wikipedia.org/wiki/Latin-script_alphabet
- https://en.wikipedia.org/wiki/List_of_languages_by_writing_system

Trello:
- https://trello.com/c/u1FYvrIZ/2241-5-allow-primarylocale-of-consultations-to-be-changed
- https://trello.com/c/3MrqbQae/2200-5-allow-primarylocale-of-documentcollections-to-be-changed