Use `#to_sym` instead of `#intern`
Although these two methods are aliases, `#to_sym` is clearer and more
explicit. It is also extremely more used in our repos: a quick GitHub
search showed that `#intern` is only used in two non-archived projects
once or twice, and both projects are Verify's, not GOV.UK's.

See the docs:
- [to_sym](https://ruby-doc.org/core-2.6/String.html#method-i-to_sym)
- [intern](https://ruby-doc.org/core-2.6/String.html#method-i-intern)