Fix flaky test in Admin::TaggableContentHelperTest
In 3eff3202e8b639e0be58077e5f0608b9b00096ab, we changed the ordering of
ministerial role appointments so they are always in descending order.

This flaky test used the default "1.day.ago" for the ministerial start
date, which meant the two ministers would sometimes be in the wrong
order by a matter of seconds.

Specifying a start date for each minister so they are always returned in
the correct order.