Add test for changing HTML attachment slugs after publishing
It appears we have long standing logic (added in
c45922dc03e4da71409a058118fba424a5db69e7) that will do the following:
- if a document has never been published before, the URL slug will be
  updated automatically each time the attachment title is changed
- if a document has previously been published, the URL slug for new (or
  modified existing) attachments will never be updated after the title
  has been entered for the first time

This behaviour was added to ensure links to previously published HTML
attachments continue functioning even if the attachment's title is changed
in a subsequent edition.

However, this causes confusion if a publisher adds a new attachment in a
later edition and finds the slug does not get updated (e.g. from a
placeholder). Therefore we need to handle both scenarios correctly.

Also updates descriptions of other related tests to ensure they reflect
what they are actually testing (the reslugging is determined on whether
the document has been published, not the edition).