Add a safely_resluggable column to attachments
This will be set to true only for HTML attachments that have never
previously been published. It will flag whether we can safely reslug a
HTML attachment when the title changes prior to publication.

Once an attachment has been published, it is no longer safely
resluggable otherwise links to that attachment will break.

We will set the value for all existing attachments to false, to ensure
the existing behaviour is maintained for attachments that have been
created.

The reason we need this column is that when a document has a new edition
created, the attachments get duplicated (with a new created_at time,
etc) and there is no link back to the previous version of that
attachment so there is no other way of telling whether an attachment has
previously been published or not.