Remove more unused finder logic
Whitehall finders (statisitics, annoucements and publications) now only
serve as foreign locale finders as from their controllers they're
redirected to Finder Frontend if the users locale is English. This means
we can remove a lot of the custom filter logic and the supporting
classes that was previously used which simplifys the code.