Allow linking to judiciary.uk
This follows on from the Publishing API change
(https://github.com/alphagov/publishing-api/pull/1897) which adds
support for this kind of redirect.

It's been approved by policy and engagement that documents can be
redirected to a domain under judiciary.uk.

See https://govuk.zendesk.com/agent/tickets/4440152 for the
background. A user wanted to unpublish some guidance and have it
redirect to appropriate replacements hosted on judiciary.uk.